requirements.txt
pyautogui>=0.9.54
pytesseract>=0.3.10
Pillow>=10.0.0
deep-translator>=1.11.0
langdetect>=1.0.9
